Rename folder and components with `V` prefix
