+++ src/components/VMediaInfo/VCopyLicense.vue
