+++ src/components/VMediaInfo/VMediaLicense.vue
